backend/requirements.txt
# FastAPI y servidor
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Base de datos
# Base de datos
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Validación y settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Autenticación
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2

# Utilidades
python-dotenv>=1.0.0
requests>=2.31.0
email-validator>=2.1.0


